+++ sql/src/main/java/org/apache/druid/sql/calcite/planner/PDQVertexFactory.java
@@ -34,6 +34,7 @@ import org.apache.druid.query.QueryDataSource;
 import org.apache.druid.query.UnionDataSource;
 import org.apache.druid.segment.column.RowSignature;
 import org.apache.druid.sql.calcite.planner.DruidQueryGenerator.InputDesc;
+import org.apache.druid.sql.calcite.planner.querygen.Vertex;
 import org.apache.druid.sql.calcite.rel.DruidQuery;
 import org.apache.druid.sql.calcite.rel.PartialDruidQuery;
 import org.apache.druid.sql.calcite.rel.PartialDruidQuery.Stage;
