+++ sql/src/main/java/org/apache/druid/sql/calcite/rel/logical/DruidTableScan.java
@@ -33,7 +33,7 @@ import org.apache.calcite.rel.core.TableScan;
 import org.apache.calcite.rel.metadata.RelMetadataQuery;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.schema.Table;
-import org.apache.druid.sql.calcite.planner.Vertex;
+import org.apache.druid.sql.calcite.planner.querygen.Vertex;
 
 import java.util.List;
